% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassGM.R
\name{assignClimateEffect}
\alias{assignClimateEffect}
\title{assignClimateEffect}
\usage{
assignClimateEffect(subCohortData, predObj, type)
}
\arguments{
\item{subCohortData}{The LandR cohortData object}

\item{predObj}{climate prediction object}

\item{type}{predict growth or mortality}
}
\description{
Calculates climate-dependent mortality and growth proportional to
 each cohorts biomass in the pixelGroup
}
