% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassGM.R
\name{calculateClimateGrowth}
\alias{calculateClimateGrowth}
\title{CalculateClimateGrowth}
\usage{
calculateClimateGrowth(subCohortData, predObj, type)
}
\arguments{
\item{subCohortData}{The LandR cohortData object}

\item{predObj}{climate prediction object}

\item{type}{predict growth or mortality}
}
\description{
CalculateClimateGrowth
}
