% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassGM.R
\name{calculateClimateMortality}
\alias{calculateClimateMortality}
\title{Calculate climate growth}
\usage{
calculateClimateMortality(subCohortData, predObj)
}
\arguments{
\item{subCohortData}{The LandR cohortData object}

\item{predObj}{climate prediction object}
}
\description{
Calculate climate growth
}
