% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassGM.R
\name{gamlss.own}
\alias{gamlss.own}
\title{gamlss.own
the definition of the backfitting additive function, Authors: Mikis Stasinopoulos, Marco Enea}
\usage{
gamlss.own(x, y, w, xeval = NULL, ...)
}
\arguments{
\item{x}{this is a param}

\item{y}{I don't know what this is}

\item{w}{I don't know what these are}
}
\description{
gamlss.own
the definition of the backfitting additive function, Authors: Mikis Stasinopoulos, Marco Enea
}
