% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassGM.R
\name{makeClimateVariablesForGMCS}
\alias{makeClimateVariablesForGMCS}
\title{makeClimateVariablesForGMCS
Prepares climate variables for use in gmcsDataPrep module with either projected
or historical (hindcasted) climate}
\usage{
makeClimateVariablesForGMCS(
  vars,
  type = "projected",
  years = 2020:2050,
  yearType = "future_years"
)
}
\arguments{
\item{vars}{Desired climate variables}

\item{type}{Character of either "projected" (i.e. forecasting climate) or "historical" (hindcasting climate)}

\item{years}{Numeric. Years for forecasting/hindcasting}

\item{yearType}{Character of either "future_period" or "historical_period"}
}
\description{
makeClimateVariablesForGMCS
Prepares climate variables for use in gmcsDataPrep module with either projected
or historical (hindcasted) climate
}
