% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassGM.R
\name{calculateClimateEffect}
\alias{calculateClimateEffect}
\title{Predict biomass change with climate variables}
\usage{
calculateClimateEffect(
  cohortData,
  pixelGroupMap,
  cceArgs,
  gmcsGrowthLimits,
  gmcsMortLimits,
  gmcsMinAge,
  cohortDefinitionCols = c("age", "speciesCode", "pixelGroup")
)
}
\arguments{
\item{cohortData}{The LandR cohortData object}

\item{pixelGroupMap}{the pixelGroupMap needed to match cohorts with raster values}

\item{cceArgs}{a list of datasets used by the climate function}

\item{gmcsGrowthLimits}{lower and upper limits to the effect of climate on growth}

\item{gmcsMortLimits}{lower and upper limits to the effect of climate on mortality}

\item{gmcsMinAge}{minimum age for which to predict full effect of growth/mortality -
younger ages are weighted toward a null effect with decreasing age}

\item{cohortDefinitionCols}{cohortData columns that determine individual cohorts}
}
\description{
Predict biomass change with climate variables
}
