% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassGM.R
\name{calculateClimateEffect}
\alias{calculateClimateEffect}
\title{CalculateClimateEffect}
\usage{
calculateClimateEffect(
  cohortData,
  CMI,
  ATA,
  gcsModel,
  mcsModel,
  pixelGroupMap,
  CMInormal,
  gmcsGrowthLimits,
  gmcsMortLimits,
  gmcsMinAge
)
}
\arguments{
\item{cohortData}{The LandR cohortData object}

\item{CMI}{A raster with annual Cumulative Moisture Index values}

\item{ATA}{A raster with annual temperature anomaly values}

\item{gcsModel}{climate-sensitive growth mixed effect model object created by gmcsDataPrep}

\item{mcsModel}{climate-sensitive mortality mixed effect model object created by gmcsDataPrep}

\item{pixelGroupMap}{the pixelGroupMap needed to match cohorts with raster values}

\item{CMInormal}{raster of CMI normals the reference period}

\item{gmcsGrowthLimits}{lower and upper limits to the effect of climate on growth}

\item{gmcsMortLimits}{lower and upper limits to the effect of climate on mortality}

\item{gmcsMinAge}{mininmum age for which to predict growth/mortality}
}
\description{
Predict biomass change with climate variables
}
