% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassGM.R
\name{calculateClimateEffect}
\alias{calculateClimateEffect}
\title{CalculateClimateEffect}
\usage{
calculateClimateEffect(cohortData, CMD, ATA, gcsModel, mcsModel,
  pixelGroupMap)
}
\arguments{
\item{cohortData}{The LandR cohortData object}

\item{CMD}{A raster with annual Cumulative Moisture Deficit values}

\item{ATA}{A raster with annual temperature anomaly values}

\item{gcsModel}{climate-sensitive growth mixed effect model object created by gmcsDataPrep}

\item{mcsModel}{climate-sensitive mortality mixed effect model object created by gmcsDataPrep}

\item{pixelGroupMap}{the pixelGroupMap needed to match cohorts with raster values}
}
\description{
Predict biomass change with climate variables
}
