% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassGM.R
\name{calculateClimateGrowth}
\alias{calculateClimateGrowth}
\title{CalculateClimateGrowth}
\usage{
calculateClimateGrowth(cohortData, predObj)
}
\arguments{
\item{cohortData}{The LandR cohortData object}

\item{predObj}{climate prediction object}
}
\description{
CalculateClimateGrowth
}
