% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassGM.R
\name{gamlss.own}
\alias{gamlss.own}
\title{gamlss.own
the definition of the backfitting additive function, Authors: Mikis Stasinopoulos, Marco Enea}
\usage{
gamlss.own(x, y, w, xeval = NULL)
}
\arguments{
\item{x}{description missing}

\item{y}{description missing}

\item{w}{description missing}

\item{xeval}{description missing}
}
\description{
gamlss.own
the definition of the backfitting additive function, Authors: Mikis Stasinopoulos, Marco Enea
}
