% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassGM.R
\name{getModelPred}
\alias{getModelPred}
\title{getModelPred}
\usage{
getModelPred(cohortData, climateCovariates, gmcsModel, stat = "growth")
}
\arguments{
\item{cohortData}{The LandR cohortData object}

\item{climateCovariates}{data.table of mean climate covariates by pixelGroup}

\item{stat}{character vector of either "growth" or "mortality"
this is used to name the pred column and exponentiate predicted growth if applicable}
}
\description{
Join cohortData with either the growth or mortality xgboost models
This is a utility function to avoid repetition as 2 predictions are needed
for each model (under the current climate, and historical) to achieve the
climate modifier
}
\keyword{internal}
