% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandR.CS-package.R
\docType{package}
\name{LandR.CS-package}
\alias{LandR.CS}
\alias{LandR.CS-package}
\title{\code{LandR.CS} package}
\description{
Utilities for 'LandR.CS' suite of landscape simulation models.
These functions incorporate climate sensitivity into LandR processes.
}
\details{
------------------------------------------------------------------------------------------
}
\section{Package options}{


\code{LandR.CS} packages use the following \code{\link{options}} to configure behaviour:

\itemize{
  \item \code{LandR.assertions}: If \code{TRUE}, additional code checks are run
        during function calls. Default \code{FALSE}.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ianmseddy/LandR.CS}
  \item Report bugs at \url{https://github.com/ianmseddy/LandR.CS/issues}
}

}
\author{
\strong{Maintainer}: Ian Eddy \email{ian.eddy@canada.com} (\href{https://orcid.org/0000-0001-7397-2116}{ORCID})

Other contributors:
\itemize{
  \item Alex M Chubaty \email{achubaty@for-cast.ca} (\href{https://orcid.org/0000-0001-7146-8135}{ORCID}) [contributor]
  \item Her Majesty the Queen in Right of Canada, as represented by the Minister of Natural Resources Canada [copyright holder]
}

}
