% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandR.CS-package.R
\docType{package}
\name{LandR.CS-package}
\alias{LandR.CS}
\alias{LandR.CS-package}
\title{\code{LandR.CS} package}
\description{
Utilities for 'LandR.CS' suite of landscape simulation models.
These functions incorporate climate sensitivity into LandR processes.
}
\details{
------------------------------------------------------------------------------------------
}
\section{Package options}{


\code{LandR.CS} packages use the following \code{\link{options}} to configure behaviour:

\itemize{
  \item \code{LandR.assertions}: If \code{TRUE}, additional code checks are run
        during function calls. Default \code{FALSE}.
}
}

\author{
\strong{Maintainer}: Ian Eddy \email{ian.eddy@canada.ca}

}
