% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassGM.R
\name{own}
\alias{own}
\title{own
 for predicting from gamlss with no random effect}
\usage{
own(
  fixed = ~1,
  random = NULL,
  correlation = NULL,
  method = "ML",
  level = NULL,
  ...
)
}
\arguments{
\item{fixed}{the fixed terms}

\item{random}{the random terms}

\item{correlation}{this is the correlation structure?}

\item{method}{Ceres help me}

\item{level}{the marginal or conditional predictor}
}
\description{
own
 for predicting from gamlss with no random effect
}
